+++ frontend/package.json
@@ -8,20 +8,23 @@
   },
   "license": "AGPL-3.0",
   "dependencies": {
+    "@stripe/react-stripe-js": "^1.1.2",
+    "@stripe/stripe-js": "^1.9.0",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
+    "file-loader": "^6.0.0",
     "next": "^9.4.1",
     "node-sass": "^4.13.0",
     "raw-loader": "^4.0.0",
-    "file-loader": "^6.0.0",
-    "url-loader": "^4.1.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.88.2",
-    "superagent": "^5.1.3"
+    "superagent": "^5.1.3",
+    "url-loader": "^4.1.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.2"
+    "nodemon": "^2.0.2",
+    "prettier": "^2.1.1"
   }
 }
