+++ frontend/pages/buy.js
@@ -1,0 +1,77 @@
+import React from 'react'
+import Link from 'next/link'
+import superagent from 'superagent'
+import { Container, Card, Grid, Divider, Input, Button } from 'semantic-ui-react'
+import { Elements, CardElement } from '@stripe/react-stripe-js'
+import { loadStripe } from '@stripe/stripe-js'
+import path from 'path'
+
+import Head from '../components/Head'
+import TitleBar from '../components/TitleBar'
+
+import styles from './buy.module.scss'
+
+// Make sure to call `loadStripe` outside of a component’s render to avoid
+// recreating the `Stripe` object on every render.
+const stripePromise = loadStripe('pk_test_JJ1eMdKN0Hp4UFJ6kWXWO4ix00jtXzq5XG')
+
+const gitea_public_url = `${process.env.KITSPACE_GITEA_URL}/api/v1`
+
+const gitea_internal_url = 'http://gitea:3000/api/v1'
+
+export default function BuyPage({ user, _csrf }) {
+  const [remoteRepo, setRemoteRepo] = React.useState('')
+  const uid = user?.id
+  return (
+    <>
+      <Head />
+      <TitleBar route="/buy" />
+      <div>hi</div>
+      <Container>
+        <Card>
+          <Card.Content>
+            <Elements stripe={stripePromise}>
+              <CardElement
+                options={{
+                  style: {
+                    base: {
+                      fontSize: '16px',
+                      color: '#424770',
+                      '::placeholder': {
+                        color: '#aab7c4',
+                      },
+                    },
+                    invalid: {
+                      color: '#9e2146',
+                    },
+                  },
+                }}
+              />
+            </Elements>
+          </Card.Content>
+        </Card>
+      </Container>
+    </>
+  )
+}
+
+function getSession(req) {
+  let session = {}
+  if (req != null && req.session) {
+    session = req.session
+  } else if (typeof window !== 'undefined' && window.session) {
+    session = window.session
+  }
+  return session
+}
+
+BuyPage.getInitialProps = async ({ req, query }) => {
+  const session = getSession(req)
+  const cookie = req?.headers?.cookie
+  const _csrf = session._csrf
+
+  return {
+    user: session.user,
+    _csrf,
+  }
+}
